Revert "【Hackathon 6th Fundable Projects 3 No.220】logsumexp int (#64993)"
This reverts commit a484208fece64bf4ca15ea01df73a5eeba4002df.

+++ paddle/phi/infermeta/unary.h
@@ -368,7 +368,7 @@ void KthvalueInferMeta(const MetaTensor& x,
 void LogicalNotInferMeta(const MetaTensor& x, MetaTensor* out);
 
 void LogsumexpInferMeta(const MetaTensor& input,
-                        const std::vector<int>& axis,
+                        const std::vector<int64_t>& axis,
                         bool keepdim,
                         bool reduce_all,
                         MetaTensor* out);
